Updated dashboard layout
